Add files via upload
I created a Data Analysis project in python on IPL matches from the 2008 season all the way to the 2019 season as it is in the CSV named matches. Check out the code and pos
reviews\comments on it. Also if any changes are needed (which i'm pretty sure are definitely needed) then mention them as well. This is my first Data Analysis Project ( apart from the one I made in school as a school project ) Thank You!